test-runner: virtio: Use appropriate index size
vq_ready would cease to be accurate after 64k messages because the
virtio index would wrap around, but the 'old' index would not.

Bug: 178591639

+++ test-runner/include/virtio.h
@@ -173,7 +173,7 @@ struct virtq {
     size_t num_bufs;
     size_t queue_id;
     struct virtq_raw* raw;
-    size_t old_used_idx;
+    uint16_t old_used_idx;
     struct virtio_config* vio;
 };
